refactor: update UsersController, UsersControllerTestTest, WorkerGroupController
