+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/controller/WorkerGroupController.java
@@ -17,11 +17,6 @@
 
 package org.apache.dolphinscheduler.api.controller;
 
-import static org.apache.dolphinscheduler.api.enums.Status.DELETE_WORKER_GROUP_FAIL;
-import static org.apache.dolphinscheduler.api.enums.Status.QUERY_WORKER_ADDRESS_LIST_FAIL;
-import static org.apache.dolphinscheduler.api.enums.Status.QUERY_WORKER_GROUP_FAIL;
-import static org.apache.dolphinscheduler.api.enums.Status.SAVE_ERROR;
-
 import org.apache.dolphinscheduler.api.exceptions.ApiException;
 import org.apache.dolphinscheduler.api.service.WorkerGroupService;
 import org.apache.dolphinscheduler.api.utils.Result;
@@ -48,6 +43,11 @@ import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.Parameters;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.tags.Tag;
+
+import static org.apache.dolphinscheduler.api.enums.v2.ExceptionStatus.DELETE_WORKER_GROUP_FAIL;
+import static org.apache.dolphinscheduler.api.enums.v2.ExceptionStatus.QUERY_WORKER_ADDRESS_LIST_FAIL;
+import static org.apache.dolphinscheduler.api.enums.v2.ExceptionStatus.QUERY_WORKER_GROUP_FAIL;
+import static org.apache.dolphinscheduler.api.enums.v2.ExceptionStatus.SAVE_ERROR;
 
 /**
  * worker group controller
