+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/enums/v2/ExceptionStatus.java
@@ -69,6 +69,11 @@ public enum ExceptionStatus {
     UPDATE_TENANT_ERROR(10087, "update tenant error", "更新租户错误"),
     DELETE_TENANT_BY_ID_ERROR(10088, "delete tenant by id error", "删除租户错误"),
     VERIFY_OS_TENANT_CODE_ERROR(10089, "verify os tenant code error", "操作系统租户验证错误"),
+    UPDATE_USER_ERROR(10092, "update user error", "更新用户错误"),
+    USER_LIST_ERROR(10099, "user list error", "查询用户列表错误"),
+    VERIFY_USERNAME_ERROR(10100, "verify username error", "用户名验证错误"),
+    UNAUTHORIZED_USER_ERROR(10101, "unauthorized user error", "查询未授权用户错误"),
+    AUTHORIZED_USER_ERROR(10102, "authorized user error", "查询授权用户错误"),
     QUERY_TASK_INSTANCE_LOG_ERROR(10103, "view task instance log error", "查询任务实例日志错误"),
     DOWNLOAD_TASK_INSTANCE_LOG_FILE_ERROR(10104, "download task instance log file error", "下载任务日志文件错误"),
     CREATE_PROCESS_DEFINITION_ERROR(10105, "create process definition error", "创建工作流错误"),
@@ -101,8 +106,11 @@ public enum ExceptionStatus {
     QUEUE_NOT_EXIST(10128, "queue {0} not exists", "队列ID[{0}]不存在"),
     QUEUE_VALUE_EXIST(10129, "queue value {0} already exists", "队列值[{0}]已存在"),
     QUEUE_NAME_EXIST(10130, "queue name {0} already exists", "队列名称[{0}]已存在"),
+    SAVE_ERROR(10136, "save error", "保存错误"),
     PREVIEW_SCHEDULE_ERROR(10139, "preview schedule error", "预览调度配置错误"),
     PARSE_TO_CRON_EXPRESSION_ERROR(10140, "parse cron to cron expression error", "解析调度表达式错误"),
+    QUERY_WORKER_GROUP_FAIL(10146, "query worker group fail ", "查询worker分组失败"),
+    DELETE_WORKER_GROUP_FAIL(10147, "delete worker group fail ", "删除worker分组失败"),
     SWITCH_PROCESS_DEFINITION_VERSION_ERROR(10151, "Switch process definition version error", "切换工作流版本出错"),
     SWITCH_PROCESS_DEFINITION_VERSION_NOT_EXIST_PROCESS_DEFINITION_ERROR(10152,
             "Switch process definition version error: not exists process definition, [process definition id {0}]",
@@ -115,12 +123,21 @@ public enum ExceptionStatus {
     QUERY_AUTHORIZED_AND_USER_CREATED_PROJECT_ERROR(10162, "query authorized and user created project error error",
             "查询授权的和用户创建的项目错误"),
     FORCE_TASK_SUCCESS_ERROR(10165, "force task success error", "强制成功任务实例错误"),
+    QUERY_WORKER_ADDRESS_LIST_FAIL(10178, "query worker address list fail ", "查询worker地址列表失败"),
     QUERY_ALERT_GROUP_ERROR(10180, "query alert group error", "查询告警组错误"),
     CURRENT_LOGIN_USER_TENANT_NOT_EXIST(10181, "the tenant of the currently login user is not specified",
             "未指定当前登录用户的租户"),
     REVOKE_PROJECT_ERROR(10182, "revoke project error", "撤销项目授权错误"),
     QUERY_AUTHORIZED_USER(10183, "query authorized user error", "查询拥有项目权限的用户错误"),
+    CREATE_USER_ERROR(10090, "create user error", "创建用户错误"),
+    QUERY_USER_LIST_PAGING_ERROR(10091, "query user list paging error", "分页查询用户列表错误"),
     QUERY_EXECUTING_WORKFLOW_ERROR(10192, "query executing workflow error", "查询运行的工作流实例错误"),
+    DELETE_USER_BY_ID_ERROR(10093, "delete user by id error", "删除用户错误"),
+    GRANT_PROJECT_ERROR(10094, "grant project error", "授权项目错误"),
+    GRANT_RESOURCE_ERROR(10095, "grant resource error", "授权资源错误"),
+    GRANT_UDF_FUNCTION_ERROR(10096, "grant udf function error", "授权UDF函数错误"),
+    GRANT_DATASOURCE_ERROR(10097, "grant datasource error", "授权数据源错误"),
+    GET_USER_INFO_ERROR(10098, "get user info error", "获取用户信息错误"),
     TASK_WITH_DEPENDENT_ERROR(10195, "task used in other tasks", "删除被其他任务引用"),
     TASK_SAVEPOINT_ERROR(10196, "task savepoint error", "任务实例savepoint错误"),
     TASK_STOP_ERROR(10197, "task stop error", "任务实例停止错误"),
@@ -225,6 +242,7 @@ public enum ExceptionStatus {
     CREATE_K8S_NAMESPACE_ERROR(1300003, "create k8s namespace error", "创建k8s命名空间错误"),
     VERIFY_K8S_NAMESPACE_ERROR(1300007, "verify k8s and namespace error", "验证k8s命名空间信息错误"),
     DELETE_K8S_NAMESPACE_BY_ID_ERROR(1300008, "delete k8s namespace by id error", "删除命名空间错误"),
+    GRANT_K8S_NAMESPACE_ERROR(1300011, "grant namespace error", "授权资源错误"),
     QUERY_UNAUTHORIZED_NAMESPACE_ERROR(1300012, "query unauthorized namespace error", "查询未授权命名空间错误"),
     QUERY_AUTHORIZED_NAMESPACE_ERROR(1300013, "query authorized namespace error", "查询授权命名空间错误"),
     QUERY_CAN_USE_K8S_NAMESPACE_ERROR(1300018, "login user query can used namespace list error", "查询可用命名空间错误"),
